+++ rio_tiler/io/base.py
@@ -25,7 +25,7 @@ from rio_tiler.errors import (
 from rio_tiler.models import BandStatistics, ImageData, Info, PointData
 from rio_tiler.tasks import multi_arrays, multi_points, multi_values
 from rio_tiler.types import AssetInfo, BBox, Indexes
-from rio_tiler.utils import normalize_bounds
+from rio_tiler.utils import cast_to_sequence, normalize_bounds
 
 
 @attr.s
@@ -283,6 +283,10 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         """Validate asset name and construct url."""
         ...
 
+    def _get_reader(self, asset_info: AssetInfo) -> Type[BaseReader]:
+        """Get Asset Reader."""
+        return self.reader
+
     def parse_expression(self, expression: str, asset_as_band: bool = False) -> Tuple:
         """Parse rio-tiler band math expression."""
         input_assets = "|".join(self.assets)
@@ -309,8 +313,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         statistics: Optional[Sequence[Tuple[float, float]]] = None,
     ):
         """Update ImageData Statistics from AssetInfo."""
-        if isinstance(indexes, int):
-            indexes = (indexes,)
+        indexes = cast_to_sequence(indexes)
 
         if indexes is None:
             indexes = tuple(range(1, img.count + 1))
@@ -322,7 +325,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
             img.dataset_statistics = [statistics[bidx - 1] for bidx in indexes]
 
     def info(
-        self, assets: Union[Sequence[str], str] = None, **kwargs: Any
+        self,
+        assets: Optional[Union[Sequence[str], str]] = None,
+        **kwargs: Any,
     ) -> Dict[str, Info]:
         """Return metadata from multiple assets.
 
@@ -338,26 +343,27 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
                 "No `assets` option passed, will fetch info for all available assets.",
                 UserWarning,
             )
-
-        assets = assets or self.assets
-
-        if isinstance(assets, str):
-            assets = (assets,)
+        assets = cast_to_sequence(assets or self.assets)
 
         def _reader(asset: str, **kwargs: Any) -> Dict:
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     return src.info()
 
         return multi_values(assets, _reader, **kwargs)
 
     def statistics(
         self,
-        assets: Union[Sequence[str], str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
-        asset_expression: Optional[Dict[str, str]] = None,  # Expression for each asset
+        assets: Optional[Union[Sequence[str], str]] = None,
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
+        asset_expression: Optional[Dict[str, str]] = None,
         **kwargs: Any,
     ) -> Dict[str, Dict[str, BandStatistics]]:
         """Return array statistics for multiple assets.
@@ -378,23 +384,24 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
                 UserWarning,
             )
 
-        assets = assets or self.assets
-
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets or self.assets)
         asset_indexes = asset_indexes or {}
         asset_expression = asset_expression or {}
 
-        def _reader(asset: str, *args, **kwargs) -> Dict:
+        def _reader(asset: str, *args: Any, **kwargs: Any) -> Dict:
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     return src.statistics(
                         *args,
-                        indexes=asset_indexes.get(asset, kwargs.pop("indexes", None)),  # type: ignore
-                        expression=asset_expression.get(asset),  # type: ignore
+                        indexes=asset_indexes.get(asset, kwargs.pop("indexes", None)),
+                        expression=asset_expression.get(asset),
                         **kwargs,
                     )
 
@@ -402,9 +409,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
 
     def merged_statistics(
         self,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         categorical: bool = False,
         categories: Optional[List[float]] = None,
         percentiles: Optional[List[int]] = None,
@@ -436,7 +443,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
                     "No `assets` option passed, will fetch statistics for all available assets.",
                     UserWarning,
                 )
-            assets = assets or self.assets
+            assets = cast_to_sequence(assets or self.assets)
 
         data = self.preview(
             assets=assets,
@@ -457,9 +464,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         tile_x: int,
         tile_y: int,
         tile_z: int,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         asset_as_band: bool = False,
         **kwargs: Any,
     ) -> ImageData:
@@ -483,9 +490,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
                 f"Tile(x={tile_x}, y={tile_y}, z={tile_z}) is outside bounds"
             )
 
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets)
         if assets and expression:
             warnings.warn(
                 "Both expression and assets passed; expression will overwrite assets parameter.",
@@ -506,12 +511,17 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         indexes = kwargs.pop("indexes", None)
 
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
-            idx = asset_indexes.get(asset) or indexes  # type: ignore
+            idx = asset_indexes.get(asset) or indexes
 
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     data = src.tile(*args, indexes=idx, **kwargs)
 
                     self._update_statistics(
@@ -545,9 +555,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
     def part(
         self,
         bbox: BBox,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         asset_as_band: bool = False,
         **kwargs: Any,
     ) -> ImageData:
@@ -564,9 +574,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets)
         if assets and expression:
             warnings.warn(
                 "Both expression and assets passed; expression will overwrite assets parameter.",
@@ -587,12 +595,17 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         indexes = kwargs.pop("indexes", None)
 
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
-            idx = asset_indexes.get(asset) or indexes  # type: ignore
+            idx = asset_indexes.get(asset) or indexes
 
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     data = src.part(*args, indexes=idx, **kwargs)
 
                     self._update_statistics(
@@ -625,9 +638,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
 
     def preview(
         self,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         asset_as_band: bool = False,
         **kwargs: Any,
     ) -> ImageData:
@@ -643,9 +656,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets)
         if assets and expression:
             warnings.warn(
                 "Both expression and assets passed; expression will overwrite assets parameter.",
@@ -666,12 +677,17 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         indexes = kwargs.pop("indexes", None)
 
         def _reader(asset: str, **kwargs: Any) -> ImageData:
-            idx = asset_indexes.get(asset) or indexes  # type: ignore
+            idx = asset_indexes.get(asset) or indexes
 
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     data = src.preview(indexes=idx, **kwargs)
 
                     self._update_statistics(
@@ -706,9 +722,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         self,
         lon: float,
         lat: float,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         asset_as_band: bool = False,
         **kwargs: Any,
     ) -> PointData:
@@ -726,9 +742,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
             PointData
 
         """
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets)
         if assets and expression:
             warnings.warn(
                 "Both expression and assets passed; expression will overwrite assets parameter.",
@@ -748,13 +762,18 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         # We fall back to `indexes` if provided
         indexes = kwargs.pop("indexes", None)
 
-        def _reader(asset: str, *args, **kwargs: Any) -> PointData:
-            idx = asset_indexes.get(asset) or indexes  # type: ignore
+        def _reader(asset: str, *args: Any, **kwargs: Any) -> PointData:
+            idx = asset_indexes.get(asset) or indexes
 
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     data = src.point(*args, indexes=idx, **kwargs)
 
                     metadata = data.metadata or {}
@@ -782,9 +801,9 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
     def feature(
         self,
         shape: Dict,
-        assets: Union[Sequence[str], str] = None,
+        assets: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
-        asset_indexes: Optional[Dict[str, Indexes]] = None,  # Indexes for each asset
+        asset_indexes: Optional[Dict[str, Indexes]] = None,
         asset_as_band: bool = False,
         **kwargs: Any,
     ) -> ImageData:
@@ -801,9 +820,7 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(assets, str):
-            assets = (assets,)
-
+        assets = cast_to_sequence(assets)
         if assets and expression:
             warnings.warn(
                 "Both expression and assets passed; expression will overwrite assets parameter.",
@@ -824,12 +841,17 @@ class MultiBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         indexes = kwargs.pop("indexes", None)
 
         def _reader(asset: str, *args: Any, **kwargs: Any) -> ImageData:
-            idx = asset_indexes.get(asset) or indexes  # type: ignore
+            idx = asset_indexes.get(asset) or indexes
 
             asset_info = self._get_asset_info(asset)
-            url = asset_info["url"]
+            reader = self._get_reader(asset_info)
+
             with self.ctx(**asset_info.get("env", {})):
-                with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+                with reader(
+                    asset_info["url"],
+                    tms=self.tms,
+                    **self.reader_options,
+                ) as src:
                     data = src.feature(*args, indexes=idx, **kwargs)
 
                     self._update_statistics(
@@ -913,7 +935,11 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         return bands
 
-    def info(self, bands: Union[Sequence[str], str] = None, *args, **kwargs: Any) -> Info:
+    def info(
+        self,
+        bands: Optional[Union[Sequence[str], str]] = None,
+        **kwargs: Any,
+    ) -> Info:
         """Return metadata from multiple bands.
 
         Args:
@@ -929,17 +955,18 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
                 UserWarning,
             )
 
-        bands = bands or self.bands
-
-        if isinstance(bands, str):
-            bands = (bands,)
+        bands = cast_to_sequence(bands or self.bands)
 
         def _reader(band: str, **kwargs: Any) -> Info:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 return src.info()
 
-        bands_metadata = multi_values(bands, _reader, *args, **kwargs)
+        bands_metadata = multi_values(bands, _reader, **kwargs)
 
         meta = {
             "bounds": self.geographic_bounds,
@@ -965,7 +992,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
     def statistics(
         self,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         categorical: bool = False,
         categories: Optional[List[float]] = None,
@@ -996,7 +1023,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
                     "No `bands` option passed, will fetch statistics for all available bands.",
                     UserWarning,
                 )
-            bands = bands or self.bands
+            bands = cast_to_sequence(bands or self.bands)
 
         data = self.preview(
             bands=bands,
@@ -1016,7 +1043,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
         tile_x: int,
         tile_y: int,
         tile_z: int,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         **kwargs: Any,
     ) -> ImageData:
@@ -1039,9 +1066,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
                 f"Tile(x={tile_x}, y={tile_y}, z={tile_z}) is outside bounds"
             )
 
-        if isinstance(bands, str):
-            bands = (bands,)
-
+        bands = cast_to_sequence(bands)
         if bands and expression:
             warnings.warn(
                 "Both expression and bands passed; expression will overwrite bands parameter.",
@@ -1058,11 +1083,19 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(band: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 data = src.tile(*args, **kwargs)
+
                 if data.metadata:
                     data.metadata = {band: data.metadata}
-                data.band_names = [band]  # use `band` as name instead of band index
+
+                # use `band` as name instead of band index
+                data.band_names = [band]
+
                 return data
 
         img = multi_arrays(bands, _reader, tile_x, tile_y, tile_z, **kwargs)
@@ -1075,7 +1108,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
     def part(
         self,
         bbox: BBox,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         **kwargs: Any,
     ) -> ImageData:
@@ -1091,9 +1124,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(bands, str):
-            bands = (bands,)
-
+        bands = cast_to_sequence(bands)
         if bands and expression:
             warnings.warn(
                 "Both expression and bands passed; expression will overwrite bands parameter.",
@@ -1110,11 +1141,19 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(band: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 data = src.part(*args, **kwargs)
+
                 if data.metadata:
                     data.metadata = {band: data.metadata}
-                data.band_names = [band]  # use `band` as name instead of band index
+
+                # use `band` as name instead of band index
+                data.band_names = [band]
+
                 return data
 
         img = multi_arrays(bands, _reader, bbox, **kwargs)
@@ -1126,7 +1165,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
     def preview(
         self,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         **kwargs: Any,
     ) -> ImageData:
@@ -1141,9 +1180,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(bands, str):
-            bands = (bands,)
-
+        bands = cast_to_sequence(bands)
         if bands and expression:
             warnings.warn(
                 "Both expression and bands passed; expression will overwrite bands parameter.",
@@ -1160,11 +1197,19 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(band: str, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 data = src.preview(**kwargs)
+
                 if data.metadata:
                     data.metadata = {band: data.metadata}
-                data.band_names = [band]  # use `band` as name instead of band index
+
+                # use `band` as name instead of band index
+                data.band_names = [band]
+
                 return data
 
         img = multi_arrays(bands, _reader, **kwargs)
@@ -1178,7 +1223,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
         self,
         lon: float,
         lat: float,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         **kwargs: Any,
     ) -> PointData:
@@ -1195,9 +1240,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
             PointData
 
         """
-        if isinstance(bands, str):
-            bands = (bands,)
-
+        bands = cast_to_sequence(bands)
         if bands and expression:
             warnings.warn(
                 "Both expression and bands passed; expression will overwrite bands parameter.",
@@ -1212,13 +1255,21 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
                 "bands must be passed either via `expression` or `bands` options."
             )
 
-        def _reader(band: str, *args, **kwargs: Any) -> PointData:
+        def _reader(band: str, *args: Any, **kwargs: Any) -> PointData:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 data = src.point(*args, **kwargs)
+
                 if data.metadata:
                     data.metadata = {band: data.metadata}
-                data.band_names = [band]  # use `band` as name instead of band index
+
+                # use `band` as name instead of band index
+                data.band_names = [band]
+
                 return data
 
         data = multi_points(bands, _reader, lon, lat, **kwargs)
@@ -1230,7 +1281,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
     def feature(
         self,
         shape: Dict,
-        bands: Union[Sequence[str], str] = None,
+        bands: Optional[Union[Sequence[str], str]] = None,
         expression: Optional[str] = None,
         **kwargs: Any,
     ) -> ImageData:
@@ -1246,9 +1297,7 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
             rio_tiler.models.ImageData: ImageData instance with data, mask and tile spatial info.
 
         """
-        if isinstance(bands, str):
-            bands = (bands,)
-
+        bands = cast_to_sequence(bands)
         if bands and expression:
             warnings.warn(
                 "Both expression and bands passed; expression will overwrite bands parameter.",
@@ -1265,11 +1314,19 @@ class MultiBandReader(SpatialMixin, metaclass=abc.ABCMeta):
 
         def _reader(band: str, *args: Any, **kwargs: Any) -> ImageData:
             url = self._get_band_url(band)
-            with self.reader(url, tms=self.tms, **self.reader_options) as src:  # type: ignore
+            with self.reader(
+                url,
+                tms=self.tms,
+                **self.reader_options,
+            ) as src:
                 data = src.feature(*args, **kwargs)
+
                 if data.metadata:
                     data.metadata = {band: data.metadata}
-                data.band_names = [band]  # use `band` as name instead of band index
+
+                # use `band` as name instead of band index
+                data.band_names = [band]
+
                 return data
 
         img = multi_arrays(bands, _reader, shape, **kwargs)