+++ rio_tiler/types.py
@@ -59,6 +59,7 @@ class AssetInfo(TypedDict, total=False):
     """Asset Reader Options."""
 
     url: str
+    media_type: str
     env: Optional[Dict]
     metadata: Optional[Dict]
     dataset_statistics: Optional[Sequence[Tuple[float, float]]]